requirements.txt
pillow
win32gui
pyautogui
pyzbar
psutil
pywin32
PyJWT
vdf
colorama